+++ canvas/src/main/java/com/angcyo/canvas/core/component/XAxis.kt
@@ -18,7 +18,7 @@ class XAxis : BaseAxis() {
         val factor = canvasViewBox.invertMatrix.getScaleX() //如果放大了, 需要扩大的因子
         val end =
             (canvasViewBox.getContentWidth() - canvasViewBox.getTranslateX()) * factor //获取刻度结束的像素位置
-        val step = canvasViewBox.valueUnit.getGraduatedScaleGap() //刻度的间隔
+        val step = canvasViewBox.valueUnit.getGraduatedScaleGap().toFloat() //刻度的间隔
 
         val scaleX = canvasViewBox.getScaleX()
         var index = 0
@@ -38,7 +38,7 @@ class XAxis : BaseAxis() {
         var pixel = canvasViewBox.getCoordinateSystemX()
         val factor = canvasViewBox.invertMatrix.getScaleX()
         val end = (0 - canvasViewBox.getTranslateX()) * factor
-        val step = canvasViewBox.valueUnit.getGraduatedScaleGap()
+        val step = canvasViewBox.valueUnit.getGraduatedScaleGap().toFloat()
 
         val scaleX = canvasViewBox.getScaleX()
         var index = 0
