+++ vector/src/main/java/com/angcyo/gcode/GCodeCmd.kt
@@ -20,13 +20,13 @@ data class GCodeCmd(
     /**指令, 比如:G M X Y*/
     val cmd: String,
     /**指令跟随的数字*/
-    val number: Float = 0f,
+    val number: Double = 0.0,
 
     /**G指令时[number]数值需要放大的倍数, 受[G20]英寸单位 [G21]毫米单位 指令影响
      * [pixel] = [number] * [ratio] */
-    val ratio: Float = 1f,
+    val ratio: Double = 1.0,
     /**G指令时[number]对应的像素点单位*/
-    val pixel: Float = 0f
+    val pixel: Double = 0.0
 )
 
 /**大写字母A~Z*/
