+++ vector/src/main/java/com/angcyo/gcode/GCodeParseConfig.kt
@@ -7,6 +7,6 @@ package com.angcyo.gcode
  */
 data class GCodeParseConfig(
     val text: String, //需要解析的GCode文本
-    val mmRatio: Float, //G21 毫米指令对应的缩放比例
-    val inRatio: Float, //G20 英寸指令对应的缩放比例
+    val mmRatio: Double, //G21 毫米指令对应的缩放比例
+    val inRatio: Double, //G20 英寸指令对应的缩放比例
 )
